% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC&AUC.R
\name{AUC}
\alias{AUC}
\title{Binary classification error rate}
\usage{
AUC(ypred, ytest, acc = 0.01, plot = TRUE)
}
\arguments{
\item{ypred}{predicte value}

\item{ytest}{TRUE value}
}
\value{
ROC and AUC
}
\description{
Binary classification error rate by AUC
}
\examples{
\dontrun{
ypred <- runif(1000)
ytest1 <- ifelse(runif(1000) >= 0.5, 1, 0)
AUC(ypred, ytest2)
}
}
