% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression ways comparion.R
\name{lambda.GCV}
\alias{lambda.GCV}
\title{cross validation to lasso}
\usage{
lambda.GCV(x, y, lambdas, type = "BIC")
}
\arguments{
\item{x}{observed value}

\item{y}{response value}

\item{lambdas}{a pre-specified sequence of lambdas}
}
\value{
Nonzero parameter R/zero parameter zero and MSE
}
\description{
cross validation to lasso BIC/AIC/CV/GCV
}
\examples{
\dontrun{
library(MASS)
library(glmnet)
n = 20
beta <- c(3, 1.5, 0, 0, 2, 0, 0, 0)
Sigma = matrix(0, 8, 8)
for(i in 1:8)
  for(j in 1:i) {
    Sigma[i,j] = Sigma[j,i] = 0.5^(i-j)
  }
noise.s <- 3
sgn.ratio <- 5.7
signal.var <- t(beta) \%*\% Sigma \%*\% t(t(beta))
noise.var <- noise.s^2
signal.s <- as.numeric(sgn.ratio*noise.var/signal.var)
d <- length(beta)
x <- mvrnorm(n = n, mu = rep(0, d), Sigma = sqrt(signal.s)*Sigma)
epsilon <- rnorm(n)
y <- t(beta \%*\% t(x) + noise.s * epsilon)
# LASSO AIC
bestlam <- lambda.GCV(x, y, 10^(seq(-2, 1, length.out = 50)), type = "AIC")
out <- glmnet(x, y, alpha = 1, lambda = bestlam)
lasso.coef <- predict(out, type = "coefficients", s = bestlam)[2:9, ]
R1 = as.numeric(lasso.coef!=0)
lasso.pred = predict(out, s = bestlam, newx = x)
mse1 <- mean((lasso.pred - y)^2)
zeros1 <- sum(lasso.coef==0)
# LASSO BIC
bestlam <- lambda.GCV(x, y, 10^(seq(-2, 1, length.out = 50)))
out <- glmnet(x, y, alpha = 1, lambda = bestlam)
lasso.coef <- predict(out, type = "coefficients", s = bestlam)[2:9, ]
R2 = as.numeric(lasso.coef!=0)
lasso.pred = predict(out, s = bestlam, newx = x)
mse2 <- mean((lasso.pred - y)^2)
zeros2 <- sum(lasso.coef==0)
# LASSO GCV
bestlam <- lambda.GCV(x, y, 10^(seq(-2, 1, length.out = 50)), type = "GCV")
out <- glmnet(x, y, alpha = 1, lambda = bestlam)
lasso.coef <- predict(out, type = "coefficients", s = bestlam)[2:9, ]
R3 = as.numeric(lasso.coef!=0)
lasso.pred = predict(out, s = bestlam, newx = x)
mse3 <- mean((lasso.pred - y)^2)
zeros3 <- sum(lasso.coef==0)
}
}
