\name{SC19012-package}
\alias{SC19012-package}
\alias{SC19012}
\docType{package}
\title{
  R package for SC
}
\description{
  All homework of SC ,and two function ,one is Binary classification error rate by AUC,and another is cross validation to lasso BIC/AIC/CV/GCV.
}
\details{
  With this R package,we can compute Binary classification error rate by AUC,and get merit by cross validation to lasso BIC/AIC/CV/GCV.
}
\author{
Xu He, hex1996@mail.ustc.edu.cn.

Maintainer: Xu He <hex1996@mail.ustc.edu.cn>
}

\keyword{ Statistical Computing }

